+++ packages/actions/noti/src/index.js
@@ -6,4 +6,5 @@
  * @packageDocumentation
  */
 
-export * from './toast.js';
+export * from './portal.js';
+export * from './store.js';
