--- packages/actions/noti/src/toast.js
@@ -1,10 +1,0 @@
-import { writable } from 'svelte/store';
-
-/**
- *
- * @param {HTMLElement} node
- * @param {*} param
- */
-export function portal(node, param) {}
-
-// export function prepare()
